requirements.txt
atlassian-python-api
langchain-ollama
python-dotenv
langgraph
langmem
